# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentContextOut": {
        "description": "Mirrors `shared.contracts.AgentContext` field-for-field.",
        "properties": {
          "current_limit": {
            "title": "Current Limit",
            "type": "integer"
          },
          "decisions_since_clawback": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisions Since Clawback"
          },
          "decisions_since_last_change": {
            "title": "Decisions Since Last Change",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/AgentState"
          }
        },
        "required": [
          "current_limit",
          "decisions_since_last_change",
          "decisions_since_clawback",
          "state"
        ],
        "title": "AgentContextOut",
        "type": "object"
      },
...
      "AgentOut": {
        "description": "A governed agent. Backend-local resource \u2014 no `shared/` equivalent.\n\n`current_rung`/`current_limit` are always kept in sync with\n`shared.constants.rung_of`/`limit_of` (the invariant `TrustEvaluation`\ndocuments applies here too, since this is the same pair of numbers).",
        "properties": {
          "context": {
            "$ref": "#/components/schemas/AgentContextOut"
          },
          "current_limit": {
            "title": "Current Limit",
            "type": "integer"
          },
          "current_rung": {
            "title": "Current Rung",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/AgentState"
          }
        },
        "required": [
          "id",
          "name",
          "current_limit",
          "current_rung",
          "state",
          "context"
        ],
        "title": "AgentOut",
        "type": "object"
      },
      "AgentState": {
        "enum": [
          "probation",
          "active",
          "restricted",
          "suspended"
        ],
        "title": "AgentState",
        "type": "string"
      },
...
      "Direction": {
        "enum": [
          "INCREASE",
          "HOLD",
          "CLAWBACK"
        ],
        "title": "Direction",
        "type": "string"
      },
...
      "ErrorBody": {
        "description": "Every non-2xx response returns this shape as its JSON body.\n\n`code` is a short machine-readable slug (e.g. \"not_found\", \"forbidden\",\n\"validation_error\") for client branching; `message` is one sentence a human\nreads; `detail` carries whatever extra structure the specific failure needs\n(field errors, the resource id that was missing, etc.) and may be null.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_AgentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "1-indexed page number",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Items per page, as requested",
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matching records, independent of page size",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[AgentOut]",
        "type": "object"
      },
...
      "Page_PolicyVersionOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PolicyVersionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "1-indexed page number",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Items per page, as requested",
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matching records, independent of page size",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[PolicyVersionOut]",
        "type": "object"
      },
...
      "Page_TrustEvaluationOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrustEvaluationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "1-indexed page number",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Items per page, as requested",
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matching records, independent of page size",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[TrustEvaluationOut]",
        "type": "object"
      },
...
      "RecommendationOut": {
        "description": "Mirrors `shared.contracts.Recommendation` field-for-field.",
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "clamped": {
            "title": "Clamped",
            "type": "boolean"
          },
          "clamped_from": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clamped From"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "governance_mode": {
            "title": "Governance Mode",
            "type": "string"
          },
          "has_dissent": {
            "title": "Has Dissent",
            "type": "boolean"
          },
          "opinions": {
            "items": {
              "$ref": "#/components/schemas/AgentOpinionOut"
            },
            "title": "Opinions",
            "type": "array"
          },
          "proposed_limit": {
            "title": "Proposed Limit",
            "type": "integer"
          },
          "proposed_rung": {
            "title": "Proposed Rung",
            "type": "integer"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RecommendationStatus"
          },
          "trust_evaluation_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Evaluation Ref"
          }
        },
        "required": [
          "recommendation_id",
          "agent_id",
          "schema_version",
          "direction",
          "proposed_limit",
          "proposed_rung",
          "rationale",
          "opinions",
          "has_dissent",
          "confidence",
          "governance_mode",
          "status",
          "trust_evaluation_ref",
          "generated_at",
          "clamped",
          "clamped_from"
        ],
        "title": "RecommendationOut",
        "type": "object"
      },
...
      "TrustEvaluationOut": {
        "description": "Mirrors `shared.contracts.TrustEvaluation` field-for-field, plus `id`\n(see module docstring \u2014 this is the one backend-minted addition).",
        "properties": {
          "accuracy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProportionResultOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "acted_decisions": {
            "title": "Acted Decisions",
            "type": "integer"
          },
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ScoreComponentOut"
            },
            "title": "Components",
            "type": "array"
          },
          "config_fingerprint": {
            "title": "Config Fingerprint",
            "type": "string"
          },
          "critical_error_rate": {
            "title": "Critical Error Rate",
            "type": "number"
          },
          "critical_errors": {
            "title": "Critical Errors",
            "type": "integer"
          },
          "critical_errors_in_recent_window": {
            "title": "Critical Errors In Recent Window",
            "type": "integer"
          },
          "current_limit": {
            "title": "Current Limit",
            "type": "integer"
          },
          "current_rung": {
            "title": "Current Rung",
            "type": "integer"
          },
          "decisions_since_last_change": {
            "title": "Decisions Since Last Change",
            "type": "integer"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "drift": {
            "$ref": "#/components/schemas/DriftResultOut"
          },
          "eligible_for_increase": {
            "title": "Eligible For Increase",
            "type": "boolean"
          },
          "escalated_decisions": {
            "title": "Escalated Decisions",
            "type": "integer"
          },
          "evaluated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluated At"
          },
          "human_agreement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProportionResultOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "noncritical_errors": {
            "title": "Noncritical Errors",
            "type": "integer"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "recommended_limit": {
            "title": "Recommended Limit",
            "type": "integer"
          },
          "recommended_rung": {
            "title": "Recommended Rung",
            "type": "integer"
          },
          "ruled_escalations": {
            "title": "Ruled Escalations",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/AgentState"
          },
          "total_decisions": {
            "title": "Total Decisions",
            "type": "integer"
          },
          "trust_score": {
            "title": "Trust Score",
            "type": "number"
          },
          "utilization": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProportionResultOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "weights_renormalised": {
            "title": "Weights Renormalised",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "agent_id",
          "schema_version",
          "total_decisions",
          "acted_decisions",
          "escalated_decisions",
          "ruled_escalations",
          "accuracy",
          "human_agreement",
          "utilization",
          "critical_errors",
          "noncritical_errors",
          "critical_error_rate",
          "critical_errors_in_recent_window",
          "trust_score",
          "components",
          "weights_renormalised",
          "drift",
          "current_limit",
          "recommended_limit",
          "current_rung",
          "recommended_rung",
          "direction",
          "state",
          "eligible_for_increase",
          "decisions_since_last_change",
          "reason_codes",
          "evaluated_at",
          "config_fingerprint"
        ],
        "title": "TrustEvaluationOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The integration point for all four lanes: the Policy Engine, persistence and audit, the HTTP contract, and the hard ceiling. See docs/SYSTEM-EXPLAINED.md for the full design.",
    "title": "Adaptive AI Governance Platform \u2014 Backend",
    "version": "1.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agents": {
      "get": {
        "description": "List every governed agent, ordered by id.",
        "operationId": "list_agents_api_v1_agents_get",
        "parameters": [
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AgentOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Agents",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}": {
      "get": {
        "description": "Fetch one agent by id, including its current `AgentContext`.",
        "operationId": "get_agent_api_v1_agents__agent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}/policy-versions": {
      "get": {
        "description": "The agent's append-only limit history, newest first.\n\n`previous_version_id` always chains to a row that exists, except the\nvery first version for an agent.",
        "operationId": "list_policy_versions_api_v1_agents__agent_id__policy_versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PolicyVersionOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Policy Versions",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}/recommendations": {
      "post": {
        "description": "Generate a fresh governance recommendation for `agent_id`: recompute\nits `TrustEvaluation` from persisted decisions, run the governance panel\nover it (`GOVERNANCE_MODE`, default `stub`), clamp the panel's proposal to\nwhat the evidence actually supports, and persist trust evaluation,\nrecommendation, and audit entry \u2014 all in this one request's transaction\n(`app.deps.get_session`; see `app/services/governance.py`).",
        "operationId": "create_recommendation_api_v1_agents__agent_id__recommendations_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A downstream dependency (e.g. governance) could not serve this request."
          }
        },
        "summary": "Create Recommendation",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}/trust": {
      "get": {
        "description": "Evaluate `agent_id`'s real persisted decision history with the real\ntrust engine, persist the result, and return it. An agent with zero\ndecisions still gets a valid `TrustEvaluation` back \u2014 `trust_engine.evaluate`\nis designed to handle an empty history, not to be called only once one\nexists (see `app/services/trust.py`).",
        "operationId": "get_current_trust_api_v1_agents__agent_id__trust_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustEvaluationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Trust",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}/trust/history": {
      "get": {
        "description": "Every `TrustEvaluation` ever persisted for this agent, newest first \u2014\nwhat the dashboard's trust-over-time chart is built from.",
        "operationId": "list_trust_history_api_v1_agents__agent_id__trust_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TrustEvaluationOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Trust History",
        "tags": [
          "agents"
        ]
      }
    },
...
  }
}